pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["devjunhong <unique3801@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
google-cloud-storage = "^2.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
